+++ src/main/java/nc/noumea/mairie/sirh/eae/dao/IEaeCampagneActionDao.java
@@ -6,7 +6,7 @@ import nc.noumea.mairie.sirh.eae.domain.EaeCampagneAction;
 
 public interface IEaeCampagneActionDao {
 
-	public int countEaeCampagneActionToSend(Date asOfDate);
+	public long countEaeCampagneActionToSend(Date asOfDate);
 	public EaeCampagneAction getNextEaeCampagneActionToSend(Date asOfDate);
 	public int setDateMailEnvoye(EaeCampagneAction eaeCampagneAction, Date dateMailEnvoye) throws DaoException;
